(fix) 修改courseModel,修改 reviewModel
- (fix)修改courseModel: survey_url改成非必填、移動 type 的位置
- (fix)修改 reviewModel: 移除 create_datetime
